// src/app/components/model/class/Client.ts
export class ClientClass {
    clientId: number = 0;
    contactPersonName: string = '';
    companyName: string = '';
    contactNo: string = '';
    city: string = '';
    state: string = '';
    pincode: string = '';
    address: string = '';
    EmployeeStrength: number = 0;
    gstNo: string = '';
    regNo: string = '';
}